Added a Way to Import TXT Files into the Classes Page
Now rather than only being able to copy and paste or type out all of the student names the user can import a TXT file in to populate the StudentsTextBox. No other file types are accepted so all error handling should be accounted for. This is also assuming that the admin will be getting these TXT files from professors (which should already be formatted) and if there are any changes needing to be made then the user can edit within the StudentsTextBox. Possibly need to look into another fix since this is using the keyword $script: to access the text box since prior to this change it was considered as NULL when trying to fetch and update it. Don't know if it is the best practice.