Can Now Import TXT Files into the Student Text Box (Late Push)
Late push as this was already accomplished on Thursday on my branch, but now it is officially connected to the rest of the project.

Will only accept true TXT files.

Next, I am updating the UI for adding and removing student folders (class management) and implementing the necessary functionality to complete the requirements within said tab.