Can Now Add Class Folders, Student Folders, VMs, but CAN'T Connect Them to the Network(s)
Updated how the Customization is being stored because when I was storing the String "None" the VM creation wasn't recognizing it as an acceptable parameter. Now if None is selected then it stores it as $null.

Very close to having the requirement checked off. However, the VMs aren't connecting to the desired Networks I believe due to once again how the user input is being stored. Will correct that and see if it fixes the problem!