Can't Connect to Network Due to Admin Privillages?
I believe that the reason I am unable to properly set up the VMs to the Networks may be because we don't have access to the already declared Network Adapters. I may be wrong, but for now I am going to leave this as is and have Professor White test is for me in the morning on his end.

Any time I have the system print me out the available port groups the adapter I chose in the GUI is there so in theory that should match up and work, but it isn't and just throws errors.

Also, added better comments to help make the code easier to understand (hopefully).