Student VMs are Created Now Based on CWU Domain
Now, if a name doesn't match a CWU Domain name then it will create the folder but not a VM for said entity. When it is a valid CWU Domain name then the folder and VM are created while also connecting it to the network.

Now the boxes don't automatically check for the Adapters section and the Class Name is an empty text box.

Also removed all print statements and display boxes that were being used for testing purposes.